added Hole Filler, now need to implement hole finder and boundary finder for both 4 and 8 type of pixels

+++ src/NumericFunctions/Function.java
@@ -3,5 +3,5 @@ package NumericFunctions;
 import javafx.util.Pair;
 
 public interface Function {
-    double calculateResult(Pair<Double,Double>... arguments) throws Exception;
+    double calculateResult(Pair<Integer,Integer>... arguments) throws Exception;
 }
